requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.8.1
django-extensions==3.1.5
django-mysql==4.8.0
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
email-validator==1.3.0
numpy==1.22.4
Pillow==9.0.0
psycopg2==2.9.9
python-dotenv==0.21.0
pytz==2021.3
sqlparse==0.4.2
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dateutil==2.9.0
openpyxl==3.1.5
pandas==2.2.2
